#110 refactored client

+++ src/network/client-handler.h
@@ -14,6 +14,6 @@ typedef struct client_handler_s {
 
 client_handler_t *client_handler_create(sdb_database_t *db);
 void client_handler_destroy(client_handler_t *handler);
-int client_handler_process_message(client_t *client, uint8_t *data, uint32_t size, void *context);
+int client_handler_process_message(client_t *client, uint8_t *data, size_t size, void *context);
 
 #endif  // SRC_NETWORK_CLIENT_HANDLER_H_
